sgap-project/sgap_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-dotenv
python-multipart
httpx
scikit-learn
numpy
pandas
joblib
qrcode[pil]
